refactor(route): changed router folder name to route
